refactor(web): move components directory to src directory
