+++ packages/web-reporter-ui/RAMReport.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { AveragedTestCaseResult } from "@perf-profiler/types";
-import { Chart } from "./components/Chart";
+import { Chart } from "./src/components/Chart";
 import { roundToDecimal } from "./utils/roundToDecimal";
 
 export const RAMReport = ({
