+++ packages/web-reporter-ui/src/components/AccordionSectionTitle.tsx
