+++ packages/web-reporter-ui/src/components/PageBackground.tsx
