+++ packages/web-reporter-ui/src/components/SimpleTable.tsx
